Refactored bazel helpers to an internal package.
