+++ internal/jfrog/url.go
@@ -1,4 +1,4 @@
-package helpers
+package jfrog
 
 import (
 	"errors"
